Cleanup the `concatenate` CLI interface
The concatenate interface is handling couple of responsibilities,
it combines files and write to a directory, and it also includes
the available links dynamically.

This commit extract it to a separate class, and also try to clean
it up as much as possible, so everything has single responsibility
and it also makes it easier to test them out.

Another things, during the conversion process we were write it to
files and then read those again, which seems to be an unnecessary,
so it also improves it, and works on the fly and only write to a
final output file. In the future, if anyone try to write really
really huge file then there might be some memory issues but that
is very rare.

It also improves the test suite so we only write to a file only
if it is necessary, and it also re-usages some of the existing
fixture files. In the future, we might consider to completely
move away form writing to any file for test purpose.