Cleanup the `relaton yaml2xml` interface
The `yaml2xml` is dealing with quite some repetitive stuff, some
other interface is also dependent on this one, so ideally it make
sense to extract it out to it's own class and also add proper test
to ensure it functioning as expected.

For now, this commit extract it out to a simple class that expose
a public interface, and under the hood it takes care of all of the
necessary steps. Since, it's in it's own so it also makes it much
more easier to test.

Another noticeable thing, whenever we are running our test suite
then it also writes a lot of file to the file system, which also
creates some occasional problem, so this commit also changes it
so we don't write the content directly to a file for test but
capture it in a string buffer and then run test on those.