Cleanup the `relaton xml2yaml` interface
The `relaton xml2yaml` interface is dealing with quite some stuff
and looking in it closely it looks like it is pretty similar to
the `relaton yaml2xml` convertor, but the main difference is it
converts the document to yaml.

So, this commit also extract it out to a separate class, which
expose one interface to convert the content directly. The YAML
& XML convertor have quite some stuff in common, so we've also
extracted those common behavior to a base convertor class and
then customize it as necessary in each of these classes.