Cleanup the `relaton xml2html` interface
The `relaton xml2html` interface, is taking a file and then it's
related styles and templates, and then it's using a renderer to
generate the html and finally write it to a file.

But we actually do have a dedicated converter for XML, so now it
makes more sense to encapsulate those behavior there and from
the CLI all we need to do is invoke that interface and let our
converter take care of the rest.