Cleanup the `relaton yaml2html` interface
The `relaton yaml2html` interface, is taking a file and then it
is passing that to the `yaml2xml` convertor. Once it returns the
content then it's adding related styles and templates and then
it's using a renderer to generate the html content and it is also
taking care of write the file the file system.

Looking at it's responsibility it looks like a lot, but since
we already have a dedicated converter for YAML, so it should be
easier to encapsulate those behavior there and from the CLI all
we need to do is invoke that interface & let the converter take
care of the rest.