Cleanup the `relaton extract` interface
The `relaton extract` is dealing with files, and it extract the
XML from folder of Metanorma XML to a desire output directory.

Looking at this interface and also the Concatenator, looks like
they are dealing with pretty much similar structure, basically
one is dealing with concatenating files and this one is doing
the opposite, extracting files to a directory.

This commit create a generic interface, and let it handle each
of this behavior based on how that interface is called, this
simplifies both of this and also makes it easier to test.

Another notable thing here, this interface is still write to
files for test, ideally we should consider moving away from
that, but let's keep that for future improvement.