Add basic doc for public interfaces
Since, we've extracted / created couple of new classes with
couple of new public interfaces, so it might be useful have
some sort of basic documentation. In these documentation we
could describe what these new interface does and how it can
be customize.

This commit does this job and it also cleanup some of the
old stuff that we don't use anymore. This also modify some
stuff so we only load stuff in the file where we're using
those library to add clear visibility.