Add support to extract single file
Currently, the `extract` interface is expecting us to provide a
directory and then it select all of the files in that directory
and extract the relaton file from those.

But this commit, add the support so this interface will not
only works on directory but single files, and keep all other
process same as before.

Fixes: #30