Add the `relaton split` interface
This commit adds an interface that will allow us to split a relaton
collection file into multiple files. This takes the collection file
and the output directory as argument. Then it will split that file
into multiple files and write it to the output directory.

By default it usages `rxl` extension for these new files, but we
can also customize that by providing the correct one option.

```sh
relaton split Relaton-Collection-File Relaton-XML-Directory
```

Fixes: 33