Refactor relaton bibdata interface
Looking at the BibData class, seems like it's doing a lot and
lots of functionality / responsibilities are packed in this
single class which also makes this class harder to reuse.

This commit takes the first step to clean up this class, and
initially it adds a simple parser that will extract out the
attributes from an relaton xml file.