Fix document splitting in specified format
The current `relaton split` interface only can parse a RXL/XML file
and write to a variation of XML file (RXL / XML) file, but the problem
arise when we define `yaml` as output extension, then it still writes
the xml content, but usages YAML for file name only. This cases to new
problems.

So, this commit fixes this issues, so now it will check for the
output format first, and if it's different than XML then it will
parse the content to that format first and then write it to the
new file.

This interface also used to write the collection file in the output
directory, which introduced some issues like 41, so this commit also
disable that for now.

Another thing, it also updates the build_filename method to remove
all of the invalid characters from a file name.