Find / discover existing bibdata rxl file
Currently, the registry generations (eg: Calconnect) depends on the
rxl field value to represents the relaton xml file.The way we handle
is to split the existing files, then concatenate it and then split it
again, this way it would fix the relaton xml link.

But now that we have optimized that process by split interface, and
we still need an way to fix those broken link, so this commit will
look for that and if available in output directory then link those
as a `bib_rxl`, and everything will work same as before.