Fix doc type related issue for index page
In XML to HTML renderer we are using some liquid templates to render
the html page for any collection, but currently we are showing the
`doctype` attributes, which is always standard.

This commit changes this to display the correct doc type in the
listing page. Currently we don't have any specific styling for all
doc type, so we aren't leaving the css class as it is, but we might
want to adjust in the near future.

Related: https://github.com/metanorma/metanorma-cli/issues/250